server/requirements.txt
# Python dependencies for server
flask
flask-cors
digi-xbee
